refactored test_logthon.py
