+++ tests/test_logthon.py
@@ -2,30 +2,13 @@
 # -*- coding: utf-8 -*-
 
 """Tests for `logthon` package"""
-from datetime import datetime
 
 from freezegun import freeze_time
 
 from logthon import logthon as ln
+from tests.utils import compose_log
 
 Logthon = ln.Logthon()
-
-
-def compose_text_message(level, message):
-    return ln.STD_FORMAT.format(
-        timestamp=datetime.now(),
-        level=level,
-        message=message
-    )
-
-
-def compose_log(level, message):
-    text_message = compose_text_message(level, message)
-    return '{color}{message}{reset_format}'.format(
-        color=ln.LOG_LEVELS[level],
-        message=text_message,
-        reset_format=ln.RESET_FORMAT
-    )
 
 
 @freeze_time('2012-01-14')
@@ -65,7 +48,7 @@ def test_critical_level(capsys):
     message = 'This is a critical test log'
     Logthon.critical(message)
     captured = capsys.readouterr()
-    assert compose_log(ln.CRITCAL_LEVEL, message) in captured.out
+    assert compose_log(ln.CRITICAL_LEVEL, message) in captured.out
 
 
 @freeze_time('2012-01-14')
